refactor: Rename main store and fix store always restoring
