+++ components/tutorial/_tutorialView.vue
@@ -22,7 +22,7 @@ import { mapActions, mapState } from 'pinia'
 import { markRaw } from 'vue'
 import tutorialOnboarding from './tutorialOnboarding.vue'
 import { useTutorials } from '~~/stores/tutorials'
-import { useMain, flags } from '~~/stores/index'
+import { useMain, flags } from '~~/stores/main'
 
 export default {
   data () {
