+++ stores/main.ts
